function for numbers added
