+++ main.h
@@ -11,5 +11,8 @@ typedef struct type
 	char letter;
 	int (*p)(va_list);
 } type_t;
+int print_c(va_list);
+int print_s(va_list);
+int print_num(va_list);
 
 #endif
